backend/requirements.txt
django==5.0.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.9.0
psycopg2-binary
python-dotenv
gunicorn
